+++ cf_pandas/__init__.py
@@ -7,7 +7,7 @@ from pkg_resources import DistributionNotFound, get_distribution
 from .accessor import CFAccessor  # noqa
 from .options import set_options  # noqa
 from .reg import Reg
-from .utils import always_iterable, astype, match_criteria_key
+from .utils import always_iterable, astype, match_criteria_key, standard_names
 from .vocab import Vocab
 from .widget import Selector, dropdown
